ports: Move FreePort to postgresql-testing and add a test case.
