+++ libs-scala/postgresql-testing/BUILD.bazel
@@ -4,6 +4,7 @@
 load(
     "//bazel_tools:scala.bzl",
     "da_scala_library",
+    "da_scala_test",
 )
 
 da_scala_library(
@@ -18,10 +19,17 @@ da_scala_library(
     deps = [
         "//bazel_tools/runfiles:scala_runfiles",
         "//ledger-api/testing-utils",
-        "//libs-scala/ports",
         "//libs-scala/resources",
         "@maven//:commons_io_commons_io",
         "@maven//:org_scalactic_scalactic_2_12",
         "@maven//:org_scalatest_scalatest_2_12",
+    ],
+)
+
+da_scala_test(
+    name = "postgresql-testing-tests",
+    srcs = glob(["src/test/suite/scala/**/*.scala"]),
+    deps = [
+        ":postgresql-testing",
     ],
 )
