+++ libs-scala/postgresql-testing/src/main/scala/com/digitalasset/testing/postgresql/FreePort.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.ports
+package com.digitalasset.testing.postgresql
 
 import java.net.{InetAddress, ServerSocket}
